Update to OpenSTV-1.7 source code as received from Eric Seidlitz, who kindly dug up OpenSTV-1.7.tar.gz from a dark corner of his computer.
This commit probably adds some things that weren't actually part of
the pristine OpenSTV-1.7 release.  The build/ directory is likely an
artifact of a Python build process Eric ran at some point, as are any
*.pyc files.  The directory "MLB 2011" is probably data from a test
Eric ran.  Most likely setup2.py is just the new setup.py.  All of
this stuff will be cleaned out in the next commit, but I wanted it in
the history in case I guess wrong about anything.